+++ massiv/CHANGELOG.md
@@ -3,9 +3,10 @@
 * Use the the new stateful workers feature of `schdeuler-1.4.0`
 * Addition of:
   * `randomArrayS`
-  * `randomArrayIO`
-  * `generateArrayStateful`
-  * `generateArrayLinearStateful`
+  * `randomArrayWS`
+  * `generateArrayWS`
+  * `generateArrayLinearWS`
+  * `mapWS`, `forWS`, `imapWS` and `iforWS`
   * and `splitLinearlyWithStatefulM_`
 
 # 0.3.3